requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
google-cloud-aiplatform==1.38.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
requests==2.31.0
python-dotenv==1.0.0 
pillow==11.3.0
dnspython==2.7.0
djangorestframework-simplejwt==5.5.1
pydantic==2.5.0
python-pptx==0.6.21
django-ratelimit==4.1.0
django-extensions==3.2.3
gunicorn==21.2.0
uvicorn[standard]==0.24.0
psutil==5.9.6